requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
motor
certifi